requirements.txt
asyncio==3.4.3
binance-connector==1.16.0
discord.py==1.7.3
requests
websocket-client
